Added support for more complex filenames
This adds support for a separator after the id, with the rest of the filename being the title.
It also adds a settings tab to specify the separator, whether the title should be added automatically, and how the id should be extracted from the title.
ID extraction supports 3 modes:
- Strict; the filename must only be the ID. Same as it currently works.
- Separator; the filename must be the id, then the separator, then the title.
- Fuzzy; treat the first non-alphanumeric character as the separator, the rest as the title.

If this is set to 'separator' you have to specify the separator.
There is also a setting to use the specified separator to automatically insert the title into the filname.